+++ src/viriatum_commons/system/socket.h
@@ -73,6 +73,7 @@
 #define SOCKET_CONNECT(socket_handle, socket_address) connect(socket_handle, (LPSOCKADDR) & socket_address, sizeof(SOCKET_ADDRESS))
 #define SOCKET_CONNECT_SIZE(socket_handle, socket_address, socket_address_size) connect(socket_handle, (LPSOCKADDR) & socket_address, socket_address_size)
 #define SOCKET_ACCEPT(socket_handle, socket_address, socket_address_size) accept(socket_handle, (LPSOCKADDR) socket_address, &socket_address_size)
+#define SOCKET_GET_NAME(socket_handle, socket_address, socket_address_size) getsockname(socket_handle, (LPSOCKADDR) & socket_address, &socket_address_size)
 #define SOCKET_CLOSE(socket_handle) closesocket(socket_handle)
 #define SOCKET_ADDRESS_CREATE(socket_address, type, address, port) \
     memset(&socket_address, 0, sizeof(SOCKET_ADDRESS));            \
@@ -155,6 +156,7 @@
 #define SOCKET_CONNECT(socket_handle, socket_address) connect(socket_handle, (struct sockaddr *) &socket_address, sizeof(SOCKET_ADDRESS))
 #define SOCKET_CONNECT_SIZE(socket_handle, socket_address, socket_address_size) connect(socket_handle, (struct sockaddr *) &socket_address, socket_address_size)
 #define SOCKET_ACCEPT(socket_handle, socket_address, socket_address_size) accept(socket_handle, (struct sockaddr *) socket_address, &socket_address_size)
+#define SOCKET_GET_NAME(socket_handle, socket_address, socket_address_size) getsockname(socket_handle, (struct sockaddr *) &socket_address, &socket_address_size)
 #define SOCKET_CLOSE(socket_handle) close(socket_handle)
 #define SOCKET_ADDRESS_CREATE(socket_address, type, address, port) \
     memset(&socket_address, 0, sizeof(SOCKET_ADDRESS));            \